linked lists

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int count = 0; /* initialize node count */
+	size_t count = 0; /* initialize node count */
 
 	if (h == NULL)
 	{
